+++ pkg/hostmeta/creation_time_other.go
@@ -1,0 +1,12 @@
+//go:build !darwin
+
+package hostmeta
+
+import (
+	"os"
+	"time"
+)
+
+func setCreationTime(_ *os.File, _ time.Time) error {
+	return ErrCreationTimeUnsupported
+}